#42 Refactor: HomeView MVVM 패턴 적용하기
