+++ KAUPark/app/src/main/java/com/example/kaupark/model/ParkingRecord.kt
@@ -1,0 +1,9 @@
+package com.example.kaupark.model
+
+import kotlin.time.Duration
+
+data class ParkingRecord(
+    val entryTime: Long? = null,
+    val exitTime: Long? = null,
+    val duration: Long? = null
+)
