+++ KAUPark/app/src/main/java/com/example/kaupark/view/fragment/ChattingList.kt
@@ -1,4 +1,4 @@
-package com.example.kaupark.fragment
+package com.example.kaupark.view.fragment
 
 import android.os.Bundle
 import android.util.Log
