+++ KAUPark/app/src/main/java/com/example/kaupark/view/fragment/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.example.kaupark.fragment
+package com.example.kaupark.view.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
