+++ KAUPark/app/src/main/java/com/example/kaupark/view/fragment/ParkingAvailable.kt
@@ -1,13 +1,15 @@
-package com.example.kaupark.fragment
+package com.example.kaupark.view.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.kaupark.databinding.ParkingTimeLeftBinding
+import com.example.kaupark.databinding.ParkingAvailableBinding
 
-class ParkingTimeLeft : Fragment() {
+
+class ParkingAvailable : Fragment() {
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -16,7 +18,6 @@ class ParkingTimeLeft : Fragment() {
             param1 = it.getString(ARG_PARAM1)
             param2 = it.getString(ARG_PARAM2)
         }
-
          */
     }
 
@@ -24,21 +25,31 @@ class ParkingTimeLeft : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val binding = ParkingTimeLeftBinding.inflate(inflater, container, false)
+        val binding = ParkingAvailableBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     /*
     companion object {
+        /**
+         * Use this factory method to create a new instance of
+         * this fragment using the provided parameters.
+         *
+         * @param param1 Parameter 1.
+         * @param param2 Parameter 2.
+         * @return A new instance of fragment ParkingAvailable.
+         */
+
+
+        // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ParkingTimeLeft().apply {
+            ParkingAvailable().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
                 }
             }
     }
-
-     */
+    */
 }
