+++ KAUPark/app/src/main/java/com/example/kaupark/view/fragment/ParkingTimeLeft.kt
@@ -1,15 +1,13 @@
-package com.example.kaupark.fragment
+package com.example.kaupark.view.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.kaupark.databinding.ParkingAvailableBinding
+import com.example.kaupark.databinding.ParkingTimeLeftBinding
 
-
-class ParkingAvailable : Fragment() {
-
+class ParkingTimeLeft : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -18,6 +16,7 @@ class ParkingAvailable : Fragment() {
             param1 = it.getString(ARG_PARAM1)
             param2 = it.getString(ARG_PARAM2)
         }
+
          */
     }
 
@@ -25,31 +24,21 @@ class ParkingAvailable : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val binding = ParkingAvailableBinding.inflate(inflater, container, false)
+        val binding = ParkingTimeLeftBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     /*
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment ParkingAvailable.
-         */
-
-
-        // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ParkingAvailable().apply {
+            ParkingTimeLeft().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
                 }
             }
     }
-    */
+
+     */
 }
