+++ KAUPark/app/src/main/java/com/example/kaupark/view/fragment/activity/MainActivity.kt
@@ -3,7 +3,7 @@ package com.example.kaupark
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.kaupark.databinding.ChattingListBinding
-import com.example.kaupark.fragment.ChattingList
+import com.example.kaupark.view.fragment.ChattingList
 import com.google.firebase.FirebaseApp
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
