+++ KAUPark/app/src/main/java/com/example/kaupark/viewmodel/HomeViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.kaupark.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class HomeViewModel: ViewModel() {
+
+}
