+++ KAUPark/app/src/main/res/layout/activity_bottom_navigation.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.BottomNavigationActivity">
+    tools:context=".view.fragment.activity.BottomNavigationActivity">
 
     <FrameLayout
         android:id="@+id/main_container"
