+++ KAUPark/app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout4"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.LoginFragment">
+    tools:context=".view.fragment.LoginFragment">
 
     <View
         android:id="@+id/loginView"
