+++ KAUPark/app/src/main/res/layout/chatting_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/chat_layout"
-    tools:context=".fragment.ChattingList" >
+    tools:context=".view.fragment.ChattingList" >
 
     <androidx.appcompat.widget.Toolbar
         android:id = "@+id/chat_toolbar"
