+++ KAUPark/app/src/main/res/layout/fragment_parking_map.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.ParkingMap">
+    tools:context=".view.fragment.ParkingMap">
 
     <androidx.appcompat.widget.Toolbar
         android:id = "@+id/maptoolbar"
