+++ KAUPark/app/src/main/res/layout/fragment_parking_map_sub.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout3"
     android:layout_width="match_parent"
     android:layout_height="72dp"
-    tools:context=".fragment.ParkingMapSubFragment">
+    tools:context=".view.fragment.ParkingMapSubFragment">
 
     <ImageView
         android:id="@+id/imageView2"
